Update run snippers action
