+++ .github/workflows/run-docs-snippets.yml
@@ -31,4 +31,4 @@ jobs:
             export INDUCTIVA_API_KEY=${{ secrets.API_DEV_TEST_KEY }}
             export INDUCTIVA_API_URL=https://api-dev.inductiva.ai
             fi
-            pytest -n auto --markdown-docs --confcutdir=docs/ docs/simulators/CaNS.md docs/simulators/FDS.md docs/simulators/SWASH.md
+            pytest -n auto --markdown-docs --confcutdir=tutorials/ tutorials/simulators/CaNS.md tutorials/simulators/FDS.md tutorials/simulators/SWASH.md
